refactor: move toCachedNamedNode (private)
